created header component and redirect after login to pricing
