+++ app/pricing/page.tsx
@@ -1,8 +1,10 @@
+import Header from "@/components/component/header";
 import Pricing from "@/components/component/pricing";
 
 export default function page() {
   return (
     <div>
+      <Header />
       <Pricing />
     </div>
   );
